requirements.txt
django_extensions==2.2.9
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
demjson==2.2.4
dj-database-url==0.5.0
Django==3.0.5
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-cron==0.5.1
django-crontab==0.7.1
django-heroku==0.3.1
djangorestframework==3.12.1
psycopg2-binary>=2.8
djongo==1.3.3
gunicorn==20.0.4
idna==2.10
imutils==0.5.3
jproperties==2.1.0
kiwisolver==1.3.0
matplotlib==3.3.2
mysqlclient==2.0.1
numpy==1.19.2
Pillow==8.0.1
psycopg2==2.8.6
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.11
whitenoise==5.2.0
django-mysql==2.2.0
django-cors-headers
